More user test cases
Create new non-auth test cases
Move anything with an auth header to UserAuthUserTestCases
Add additional user-auth test cases

+++ mail_safe_test/resources/user.py
@@ -120,5 +120,7 @@ class UserAPI(Resource):
     @user_required
     def delete(self):
         user = current_user()
-        user.key.delete()  # Delete a single user.
+        if not user:
+            abort(400)
+        user.key.delete() # Delete a single user.
         return make_response("", 204)
